requirements.txt
pytest~=9.0.2
Flask~=3.1.2
PyJWT~=2.10.1
SQLAlchemy~=2.0.45
Werkzeug~=3.1.4
flask>=2.0.0
flask-cors>=3.0.0
flask-sqlalchemy>=2.5.0
pytest>=6.0.0
pytest-mock>=3.6.0
python-decouple>=3.6